Updated to include new UI and secondary scoring options.
Now with string values instead of hardcoded values.

Replaced "null" values returned for secondary UI switching to "view"

Defined all the string variables.

Removed all the magic numbers with the exception of all values that are either dp or sp.

Adjusted the layout so that buttons appear to replace each other, depending on the score type you are tallying.